Atualização na camada de api
Foi retirado a classe main e foi criado duas novas classe sobre rotas de acesso, através disso vai gerar mais fácil manutenção e deixar o código mais legível
