+++ src/api/__init__.py
@@ -3,4 +3,5 @@ from flask import Flask, request
 app = Flask(__name__)
 request_data = request
 
-import src.api.main
+import src.api.bhaskara_route
+import src.api.healthcheck_route
