+++ src/api/bhaskara_route.py
@@ -1,16 +1,6 @@
-from datetime import date
 from uptime import uptime
 from src.api import app, request_data
 from src.domain.handlers.bhaskara_handler import bhaskara_handle
-
-
-@app.route("/1/healthcheck", methods=['GET'])
-async def read_root():
-    return {
-        'uptime': uptime(),
-        'message': 'Ok',
-        'date': date.today()
-    }
 
 
 @app.route('/1/bhaskara', methods=['POST'])
